// frontend/src/types/Employee.ts
export type Employee = {
  employeeId: string
  employeeName: string
  region: string
  country: string
  city: string
  timezone: string
  department: string
  discipline: string
  roleArchetype: string
  grade: string
  careerLevel: number
  primaryDomain: string
  secondaryDomain: string
  availabilityCategory: string
  currentAllocationFTE: number
  availableFTECurrent: number
  expectedReleaseDate: string
  releaseWindow: string
  ewaStatus: string
  currentAccountId: string
  currentProjectId: string
  currentRole: string
  currentProjectStart: string
  currentProjectEnd: string
  workMode: string
  profileId: string
}